AdventOfCode.Tests/Puzzles/Puzzle25Tests.cs
namespace AdventOfCode.Tests.Puzzles;

public class Puzzle25Tests
{
    private Puzzle25 _puzzle;

    public Puzzle25Tests()
    {
        _puzzle = new Puzzle25();
    }

    [Fact]
    public void SolvePart1()
    {
        var result = _puzzle.SolvePart1();
        result.Should().Be(2770);
    }

    [Fact(Skip = "Not yet implemented")]
    public void SolvePart2()
    {
        var result = _puzzle.SolvePart2();
        result.Should().Be(0);
    }

    [Fact]
    public void TestPart1Example()
    {
        _puzzle = new Puzzle25(
            "#####",
            ".####",
            ".####",
            ".####",
            ".#.#.",
            ".#...",
            ".....",
            "",
            "#####",
            "##.##",
            ".#.##",
            "...##",
            "...#.",
            "...#.",
            ".....",
            "",
            ".....",
            "#....",
            "#....",
            "#...#",
            "#.#.#",
            "#.###",
            "#####",
            "",
            ".....",
            ".....",
            "#.#..",
            "###..",
            "###.#",
            "###.#",
            "#####",
            "",
            ".....",
            ".....",
            ".....",
            "#....",
            "#.#..",
            "#.#.#",
            "#####"
        );
        var result = _puzzle.SolvePart1();
        result.Should().Be(3);
    }
}
